ref: clean up the new structure of fixtures
