+++ tests/fixtures/aula-tests-fixture.ts
@@ -1,0 +1,148 @@
+import { test as baseTest, BrowserContext, Page } from '@playwright/test';
+import fs from 'fs';
+import * as userInteractions from '../interactions/users';
+import { RoomData, UserData } from '../support/types';
+import * as shared from '../support/utils';
+import { FILTER_EXCLUDED_RESOURCES, getStorageStatePath } from './utils';
+import { DbBackchannel } from './db-backchannel';
+
+type WorkerFixtures = {
+  // Fixture to determine the DB instance_code based on worker's parallelIndex
+  dbInstanceCode: string;
+
+  // Fixture to ensure singleton DbBackchannel is loaded and DB is re-seeded
+  baselineLoaded: void;
+
+  // Fixture that ensures browser context is bootstrapped, returning local file path of the context
+  ensureStatePathFor: (username: 'student' | 'user' | 'admin') => Promise<string>;
+
+  // Convenience fixtures to return value objects representing data seeded to the DB
+  seededRoom: RoomData;
+  seededStudent: UserData;
+  seededUser: UserData;
+};
+
+type BrowserFixtures = {
+  newContextFor: (role: 'student' | 'user' | 'admin') => Promise<BrowserContext>;
+  newPageFor: (role: 'student' | 'user' | 'admin') => Promise<Page>;
+};
+
+export const test = baseTest.extend<BrowserFixtures, WorkerFixtures>({
+  /**
+   * Prefer `parallelIndex` over `workerIndex` because it's always in the stable range of
+   * 0..${workers - 1}, while `workerIndex` is always incremented, beyond max workers count
+   */
+  dbInstanceCode: [async ({ }, use, { parallelIndex }) => {
+    const idx = parallelIndex ?? '0';
+    const name = `db00${idx}`;
+    await use(name);
+  }, { scope: 'worker' }],
+
+  baselineLoaded: [async ({ dbInstanceCode }, use) => {
+    const dbBackchannel = await DbBackchannel.getByInstanceCode(dbInstanceCode);
+    await dbBackchannel.truncateAll();
+    await dbBackchannel.seed();
+    await use(undefined);
+  }, { scope: 'worker' }],
+
+  ensureStatePathFor: [async ({ baselineLoaded, dbInstanceCode, browser }, use) => {
+    const factory = async (username: 'student' | 'user' | 'admin') => {
+      const storageStatePath = getStorageStatePath(`${dbInstanceCode}_${username}`);
+
+      // use browser to create state if missing
+      if (!fs.existsSync(storageStatePath)) {
+        const ctx = await browser.newContext();
+        const page = await ctx.newPage();
+        await page.route('**/*', FILTER_EXCLUDED_RESOURCES);
+
+        console.log(`↔️ Accessing tested instance ${dbInstanceCode} as ${username}`);
+        await page.goto(shared.getHost(), { waitUntil: 'domcontentloaded' });
+        await userInteractions.ensureSpecificInstanceEntered(page, dbInstanceCode);
+        await userInteractions.login(page, { username: username, password: 'aula' });
+        await page.waitForFunction(() => localStorage.getItem('token'));
+
+        await ctx.storageState({ path: storageStatePath });
+        await ctx.close();
+      }
+      return storageStatePath;
+    }
+
+    return await use(factory);
+
+    // @NOTE: no cleanup happening here, we let the global teardown cleanup all leftover auth state files
+  }, { scope: 'worker' }],
+
+  seededStudent: [async ({ dbInstanceCode, baselineLoaded }, use) => {
+    // return minimal descriptor from baseline; no DB mutation
+    await use({
+      displayName: 'e2e.student',
+      realName: 'Student Studentić',
+      username: 'student',
+      hashId: 'e2e.user.20.student',
+      role: 20,
+      about: 'e2e test student',
+      password: 'aula',
+      email: 'dev+e2e-student@aula.de'
+    });
+  }, { scope: 'worker' }],
+
+  seededUser: [async ({ dbInstanceCode, baselineLoaded }, use) => {
+    // return minimal descriptor from baseline; no DB mutation
+    await use({
+      displayName: 'e2e.user',
+      realName: 'User Userić',
+      username: 'user',
+      hashId: 'e2e.user.20.user',
+      role: 20,
+      about: 'e2e test user',
+      password: 'aula',
+      email: 'dev+e2e-user@aula.de'
+    });
+  }, { scope: 'worker' }],
+
+  seededRoom: [async ({ dbInstanceCode, baselineLoaded }, use) => {
+    // return minimal descriptor from baseline; no DB mutation
+    await use({
+      name: 'e2e.class_1A',
+      description: 'e2e test room of class 1A',
+      users: [{ username: 'user' }, { username: 'student' }]
+    });
+  }, { scope: 'worker' }],
+
+  // factory to create a new authenticated context (test-scoped)
+  newContextFor: [async ({ browser, ensureStatePathFor }, use) => {
+    const [student, user, admin] = await Promise.all([
+      ensureStatePathFor('student'),
+      ensureStatePathFor('user'),
+      ensureStatePathFor('admin'),
+    ]);
+    const mapping: Record<string, string> = { student, user, admin };
+    // track created contexts for cleanup
+    const created: BrowserContext[] = [];
+    const factory = async (username: 'student' | 'user' | 'admin') => {
+      const state = mapping[username];
+      const ctx = await browser.newContext({ storageState: state });
+      created.push(ctx);
+      return ctx;
+    };
+    await use(factory);
+    // cleanup after test
+    await Promise.all(created.map(c => c.close().catch(() => { })));
+  }, { scope: 'test' }],
+
+  // factory to create a fresh page for a role
+  newPageFor: [async ({ newContextFor }, use) => {
+    const createdPages: Page[] = [];
+    const factory = async (role: 'student' | 'user' | 'admin') => {
+      const ctx = await newContextFor(role);
+      const page = await ctx.newPage();
+      createdPages.push(page);
+      return page;
+    };
+    await use(factory);
+    // ensure pages closed after test is executed (contexts will be closed by newContextFor cleanup)
+    await Promise.all(createdPages.map(p => p.close().catch(() => { })));
+  }, { scope: 'test' }],
+});
+
+export { expect } from '@playwright/test';
