+++ tests/fixtures/sql/100-truncate-all.sql
