--- tests/lifecycle/db-backchannel/teardown-auth.ts
@@ -1,7 +1,0 @@
-import { cleanupAuthStates } from './cleanup';
-
-export default async function globalTeardown() {
-  console.log('🧹 Cleaning up auth states...');
-
-  await cleanupAuthStates();
-}
